Support for listeing to section updates.
