+++ lib/components/page.dart
@@ -1,33 +1,18 @@
-import 'package:auto_route/auto_route.dart';
-import 'package:brunstadtv_app/components/achievement_section.dart';
 import 'package:brunstadtv_app/components/error_generic.dart';
 import 'package:brunstadtv_app/components/loading_generic.dart';
-import 'package:brunstadtv_app/components/see_more.dart';
 import 'package:brunstadtv_app/graphql/client.dart';
 import 'package:brunstadtv_app/graphql/queries/sections.graphql.dart';
 import 'package:brunstadtv_app/models/analytics/sections.dart';
-import 'package:brunstadtv_app/router/router.gr.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../graphql/queries/page.graphql.dart';
 import '../helpers/extensions.dart';
 import '../helpers/ui/sections.dart';
 import '../models/pagination_status.dart';
 import '../providers/inherited_data.dart';
-import 'card_section.dart';
-import 'featured_section.dart';
-import 'message_section.dart';
-import 'page_details_section.dart';
-import 'icon_section.dart';
-import 'label_section.dart';
-import 'sections/thumbnail_grid/item_section_thumbnail_grid.dart';
-import 'sections/thumbnail_slider/item_section_thumbnail_slider.dart';
-import 'icon_grid_section.dart';
-import 'sections/list_section.dart';
-import 'page_section.dart';
-import 'web_section.dart';
+import 'section_handle_update.dart';
 
 /// How close to the bottom of the page do you have to be before we load more items
 const kLoadMoreBottomScrollOffset = 300;
@@ -61,88 +46,6 @@ class _BccmPageState extends ConsumerState<BccmPage> {
   void initState() {
     super.initState();
     scrollController = widget.scrollController ?? ScrollController();
-  }
-
-  Widget? _getSectionWidget(Fragment$Section s) {
-    final extraItems = paginationMap[s.id]?.items;
-    final iconSection = s.asOrNull<Fragment$Section$$IconSection>();
-    if (iconSection != null) {
-      return PageSection.fromFragment(iconSection, child: IconSection(iconSection));
-    }
-    final labelSection = s.asOrNull<Fragment$Section$$LabelSection>();
-    if (labelSection != null) {
-      return PageSection.fromFragment(labelSection, child: LabelSection(labelSection));
-    }
-    final defaultSection = s.asOrNull<Fragment$Section$$DefaultSection>();
-    if (defaultSection != null) {
-      return PageSection.fromFragment(defaultSection, child: ItemSectionThumbnailSlider.fromDefaultSection(defaultSection));
-    }
-    final posterSection = s.asOrNull<Fragment$Section$$PosterSection>();
-    if (posterSection != null) {
-      return PageSection.fromFragment(posterSection, child: ItemSectionThumbnailSlider.fromPosterSection(posterSection));
-    }
-    var defaultGridSection = s.asOrNull<Fragment$Section$$DefaultGridSection>();
-    if (defaultGridSection != null) {
-      defaultGridSection = defaultGridSection.copyWith(
-        items: defaultGridSection.items.copyWith(items: [
-          ...defaultGridSection.items.items,
-          ...(extraItems?.whereType<Fragment$Section$$DefaultGridSection$items$items>().toList() ?? [])
-        ]),
-      );
-      return PageSection.fromFragment(defaultGridSection, child: ItemSectionThumbnailGrid.fromDefaultGridSection(defaultGridSection));
-    }
-    var posterGridSection = s.asOrNull<Fragment$Section$$PosterGridSection>();
-    if (posterGridSection != null) {
-      posterGridSection = posterGridSection.copyWith(
-        items: posterGridSection.items.copyWith(
-            items: [...posterGridSection.items.items, ...(extraItems?.whereType<Fragment$Section$$PosterGridSection$items$items>().toList() ?? [])]),
-      );
-      return PageSection.fromFragment(posterGridSection, child: ItemSectionThumbnailGrid.fromPosterGridSection(posterGridSection));
-    }
-    final featuredSection = s.asOrNull<Fragment$Section$$FeaturedSection>();
-    if (featuredSection != null) {
-      return PageSection.fromFragment(featuredSection, child: FeaturedSection(featuredSection));
-    }
-    final iconGridSection = s.asOrNull<Fragment$Section$$IconGridSection>();
-    if (iconGridSection != null) {
-      return PageSection.fromFragment(iconGridSection, child: IconGridSection(iconGridSection));
-    }
-    var listSection = s.asOrNull<Fragment$Section$$ListSection>();
-    if (listSection != null) {
-      listSection = listSection.copyWith(
-          items: listSection.items
-              .copyWith(items: [...listSection.items.items, ...(extraItems?.whereType<Fragment$Section$$ListSection$items$items>().toList() ?? [])]));
-      return PageSection.fromFragment(listSection, child: ListSection(listSection));
-    }
-    final webSection = s.asOrNull<Fragment$Section$$WebSection>();
-    if (webSection != null) {
-      return PageSection.fromFragment(webSection, child: WebSection(webSection));
-    }
-    final messageSection = s.asOrNull<Fragment$Section$$MessageSection>();
-    if (messageSection != null) {
-      return Padding(padding: const EdgeInsets.only(top: 4), child: PageSection.fromFragment(messageSection, child: MessageSection(messageSection)));
-    }
-    final cardSection = s.asOrNull<Fragment$Section$$CardSection>();
-    if (cardSection != null && cardSection.items.items.isNotEmpty) {
-      return Padding(padding: const EdgeInsets.only(top: 4), child: PageSection.fromFragment(cardSection, child: CardSection(cardSection)));
-    }
-    final pageDetailsSection = s.asOrNull<Fragment$Section$$PageDetailsSection>();
-    if (pageDetailsSection != null) {
-      return Padding(padding: const EdgeInsets.only(top: 4), child: PageDetailsSection(pageDetailsSection));
-    }
-    final achievementsSection = s.asOrNull<Fragment$Section$$AchievementSection>();
-    if (achievementsSection != null) {
-      return Padding(
-        padding: const EdgeInsets.only(top: 4),
-        child: PageSection(
-          title: achievementsSection.title,
-          description: achievementsSection.description,
-          rightSlot: SeeMoreSlot(onTap: () => context.router.navigate(const AchievementsScreenRoute())),
-          child: AchievementSection(achievementsSection),
-        ),
-      );
-    }
-    return null;
   }
 
   Widget getPage(context, Query$Page$page pageData) {
@@ -179,14 +82,10 @@ class _BccmPageState extends ConsumerState<BccmPage> {
             itemCount: sectionItems.length,
             itemBuilder: ((context, index) {
               var s = sectionItems[index];
-              final widget = _getSectionWidget(s);
-              if (widget != null) {
-                return InheritedData<SectionAnalytics>(
-                  inheritedData: SectionAnalytics(id: s.id, position: index, type: s.$__typename, name: s.title),
-                  child: (context) => widget,
-                );
-              }
-              return const SizedBox.shrink();
+              return InheritedData<SectionAnalytics>(
+                inheritedData: SectionAnalytics(id: s.id, position: index, type: s.$__typename, name: s.title),
+                child: (context) => SectionHandleUpdates(section: s, extraItems: paginationMap[s.id]?.items),
+              );
             }),
           ),
         ),
